% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get the file path to the raw_mandates.csv, and county_mandates.csv}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{name of the file in quotes}
}
\description{
USMaskMandates comes with 2 data sets: raw_mandates and county_mandates located in the `inst/extdata` directory.
This funcion allows for easy access.
}
\details{
The data in each file is the same as \code{\link{raw_mandates}} and \code{\link{county_mandates}}
}
