% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USMaskMandates.R
\docType{package}
\name{USMaskMandates-package}
\alias{USMaskMandates}
\alias{USMaskMandates-package}
\title{USMaskMandates: Dates of Implemenation of Mask Mandtes in US During COVID19}
\description{
A novel data set collected by students at the Univeristy of Chicago under the direction of of Austin L. Wright.
    The data is of county level data and was released with the working paper by Austin L. Wright.
}
\keyword{internal}
