% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USMaskMandates.R
\docType{package}
\name{USMaskMandates-package}
\alias{USMaskMandates}
\alias{USMaskMandates-package}
\title{USMaskMandates: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
